Added TabPanel as a component
