+++ frontend/src/components/Auth/AuthDialog.tsx
@@ -4,37 +4,10 @@ import DialogTitle from '@mui/material/DialogTitle';
 import DialogContent from '@mui/material/DialogContent';
 import LoginForm from './LoginForm';
 import SignupForm from './SignupForm';
-import Box from '@mui/material/Box';
 import Tabs from '@mui/material/Tabs';
 import Tab from '@mui/material/Tab';
 import { useTheme } from '@mui/material';
-
-function TabPanel(props: {
-  children?: React.ReactNode;
-  index: number;
-  value: number;
-}) {
-  const { children, value, index, ...other } = props;
-
-  return (
-    <div
-      role="tabpanel"
-      hidden={value !== index}
-      id={`auth-tabpanel-${index}`}
-      aria-labelledby={`auth-tab-${index}`}
-      {...other}
-    >
-      {value === index && <Box sx={{ p: 3 }}>{children}</Box>}
-    </div>
-  );
-}
-
-function a11yProps(index: number) {
-  return {
-    id: `auth-tab-${index}`,
-    'aria-controls': `auth-tabpanel-${index}`,
-  };
-}
+import { TabPanel, a11yProps } from '../Shared/TabPanel';
 
 export default function AuthDialog({
   open, onClose
