+++ frontend/src/components/Shared/TabPanel.tsx
@@ -1,0 +1,33 @@
+import Box from '@mui/material/Box';
+
+function TabPanel(props: {
+  children?: React.ReactNode;
+  index: number;
+  value: number;
+}) {
+  const { children, value, index, ...other } = props;
+
+  return (
+    <div
+      role="tabpanel"
+      hidden={value !== index}
+      id={`auth-tabpanel-${index}`}
+      aria-labelledby={`auth-tab-${index}`}
+      {...other}
+    >
+      {value === index && <Box sx={{ p: 3 }}>{children}</Box>}
+    </div>
+  );
+}
+  
+function a11yProps(index: number) {
+  return {
+    id: `auth-tab-${index}`,
+    'aria-controls': `auth-tabpanel-${index}`,
+  };
+}
+
+export {
+  TabPanel,
+  a11yProps
+}
